+++ pyffi/qskope/__init__.py
@@ -55,6 +55,7 @@ from pyffi.formats.esp import EspFormat
 from pyffi.formats.tri import TriFormat
 from pyffi.formats.bsa import BsaFormat
 from pyffi.formats.psk import PskFormat
+from pyffi.formats.rockstar.dir_ import DirFormat
 
 from pyffi.object_models import FileFormat
 
@@ -190,7 +191,7 @@ class QSkope(QtGui.QMainWindow):
             # try reading as a nif file
             for Format in (NifFormat, CgfFormat, KfmFormat, DdsFormat,
                            TgaFormat, EgmFormat, EspFormat, TriFormat,
-                           EgtFormat, BsaFormat, PskFormat):
+                           EgtFormat, BsaFormat, PskFormat, DirFormat):
                 self.data = Format.Data()
                 try:
                     self.data.read(stream)