+++ rundoctest.py
@@ -36,6 +36,7 @@ import pyffi.formats.tri
 import pyffi.formats.bsa
 import pyffi.formats.egt
 import pyffi.formats.psk
+import pyffi.formats.rockstar.dir_
 import pyffi.spells
 import pyffi.spells.nif
 import pyffi.spells.nif.fix